+++ RELEASE_NOTES.md
@@ -4,6 +4,11 @@
 - Added support for `patterns` and `help` in the `parameters_meta` section of a WDL task. For more information, see [docs/ExpertOptions](doc/ExpertOptions.md#parameter_meta-section)
 - Upgrade to Cromwell v48
 - Upgrade to dxfuse v0.17
+- Added support for a `path` command line argument to `dxni`.
+
+```
+java -jar dxWDL-v1.41.jar dxni --path /MY_APPLETS/assemble --project project-xxxx --language wdl_v1.0 --output headers.wdl
+```
 
 ## 1.40  19-Dec-2019
 - Replaced the dxjava package with scala code. The dnanexus calls now go through the low-level DXAPI java